Unify plugin utilities and code
