+++ tests/test_plugins/test_execute_fails_plugin.cpp
@@ -3,6 +3,7 @@
 
 #include <iostream>
 
+#include "test_plugin_common.hpp"
 #include <hipdnn_sdk/data_objects/engine_details_generated.h>
 #include <hipdnn_sdk/logging/logger.hpp>
 #include <hipdnn_sdk/plugin/engine_plugin_api.h>
@@ -26,52 +27,13 @@ using namespace hipdnn_plugin;
 // NOLINTNEXTLINE(modernize-avoid-c-arrays)
 thread_local char Plugin_last_error_manager::last_error[HIPDNN_PLUGIN_ERROR_STRING_MAX_LENGTH] = "";
 
-#define LOG_API_ENTRY(format, ...) \
-    HIPDNN_LOG_INFO("API called: [{}] " format, __func__ __VA_OPT__(, ) __VA_ARGS__)
-
-#define LOG_API_SUCCESS(func_name, format, ...) \
-    HIPDNN_LOG_INFO("API success: [{}] " format, func_name __VA_OPT__(, ) __VA_ARGS__)
-
-template <typename T>
-void throw_if_null(T* value)
-{
-    if(value == nullptr)
-    {
-        throw hipdnn_plugin::Hipdnn_plugin_exception(HIPDNN_PLUGIN_STATUS_BAD_PARAM,
-                                                     std::string(typeid(T).name()) + " is nullptr");
-    }
-}
-
-class Test_container
-{
-};
-
-// Holds a weak pointer so subsequent calls to hipdnnEnginePluginCreate
-// can reuse the same instance of Test_container if it exists.
-std::weak_ptr<Test_container> test_container_lifecycle_ptr;
-
-struct hipdnnEnginePluginHandle
-{
-public:
-    virtual ~hipdnnEnginePluginHandle() = default;
-
-    hipStream_t stream = nullptr;
-
-    std::shared_ptr<Test_container> test_container;
-};
-
-struct hipdnnEnginePluginExecutionContext
-{
-    uint64_t dummy; // Placeholder
-};
-
 extern "C" {
 
 hipdnnPluginStatus_t hipdnnPluginGetName(const char** name)
 {
     LOG_API_ENTRY("name_ptr={:p}", static_cast<void*>(name));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(name);
 
         *name = plugin_name;
@@ -84,7 +46,7 @@ hipdnnPluginStatus_t hipdnnPluginGetVersion(const char** version)
 {
     LOG_API_ENTRY("version_ptr={:p}", static_cast<void*>(version));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(version);
 
         *version = plugin_version;
@@ -97,7 +59,7 @@ hipdnnPluginStatus_t hipdnnPluginGetType(hipdnnPluginType_t* type)
 {
     LOG_API_ENTRY("type_ptr={:p}", static_cast<void*>(type));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(type);
 
         *type = HIPDNN_PLUGIN_TYPE_ENGINE;
@@ -110,7 +72,7 @@ void hipdnnPluginGetLastErrorString(const char** error_str)
 {
     LOG_API_ENTRY("error_str_ptr={:p}", static_cast<void*>(error_str));
 
-    hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    try_catch([&, api_name = __func__]() {
         throw_if_null(error_str);
 
         *error_str = Plugin_last_error_manager::get_last_error();
@@ -122,7 +84,7 @@ void hipdnnPluginGetLastErrorString(const char** error_str)
 // Once plugins are loaded via plugin manager then logging will work for them
 hipdnnPluginStatus_t hipdnnPluginSetLoggingCallback(hipdnnCallback_t callback)
 {
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(callback);
         hipdnn::logging::initialize_callback_logging(plugin_name, callback);
         LOG_API_SUCCESS(api_name, "");
@@ -133,37 +95,10 @@ hipdnnPluginStatus_t hipdnnEnginePluginCreate(hipdnnEnginePluginHandle_t* handle
 {
     LOG_API_ENTRY("handle_ptr={:p}", static_cast<void*>(handle));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
 
         *handle = new hipdnnEnginePluginHandle();
-
-        (*handle)->stream = hipStreamDefault;
-
-        auto test_container_ptr = test_container_lifecycle_ptr.lock();
-        if(test_container_ptr != nullptr)
-        {
-            (*handle)->test_container = test_container_ptr;
-        }
-        else
-        {
-            static std::mutex test_container_mutex;
-            std::lock_guard<std::mutex> lock(test_container_mutex);
-
-            // if we do have a race condition that results in threads getting locked, we want to
-            // ensure that we only create one instance.  Therefore, the second thread to get
-            // through will just read from the weak pointer rather than create a new instance.
-            test_container_ptr = test_container_lifecycle_ptr.lock();
-            if(test_container_ptr != nullptr)
-            {
-                (*handle)->test_container = test_container_ptr;
-            }
-            else
-            {
-                (*handle)->test_container = std::make_shared<Test_container>();
-                test_container_lifecycle_ptr = (*handle)->test_container;
-            }
-        }
 
         LOG_API_SUCCESS(api_name, "created_handle={:p}", static_cast<void*>(*handle));
     });
@@ -173,7 +108,7 @@ hipdnnPluginStatus_t hipdnnEnginePluginDestroy(hipdnnEnginePluginHandle_t handle
 {
     LOG_API_ENTRY("handle={:p}", static_cast<void*>(handle));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
 
         delete handle;
@@ -189,10 +124,8 @@ hipdnnPluginStatus_t hipdnnEnginePluginSetStream(hipdnnEnginePluginHandle_t hand
     LOG_API_ENTRY(
         "handle={:p}, stream_id={:p}", static_cast<void*>(handle), static_cast<void*>(stream));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
-
-        handle->stream = stream;
 
         LOG_API_SUCCESS(api_name, "");
     });
@@ -212,7 +145,7 @@ hipdnnPluginStatus_t
                   max_engines,
                   static_cast<void*>(num_engines));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(op_graph);
         throw_if_null(engine_ids);
@@ -240,7 +173,7 @@ hipdnnPluginStatus_t hipdnnEnginePluginGetEngineDetails(hipdnnEnginePluginHandle
                   static_cast<const void*>(op_graph),
                   static_cast<void*>(engine_details));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(op_graph);
         throw_if_null(engine_details);
@@ -268,7 +201,7 @@ hipdnnPluginStatus_t hipdnnEnginePluginDestroyEngineDetails(hipdnnEnginePluginHa
                   static_cast<void*>(handle),
                   static_cast<void*>(engine_details));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(engine_details);
         throw_if_null(engine_details->ptr);
@@ -291,7 +224,7 @@ hipdnnPluginStatus_t
                   static_cast<const void*>(op_graph),
                   static_cast<void*>(workspace_size));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(engine_config);
         throw_if_null(op_graph);
@@ -315,7 +248,7 @@ hipdnnPluginStatus_t hipdnnEnginePluginCreateExecutionContext(
                   static_cast<const void*>(op_graph),
                   static_cast<void*>(execution_context));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(engine_config);
         throw_if_null(op_graph);
@@ -338,7 +271,7 @@ hipdnnPluginStatus_t hipdnnEnginePluginDestroyExecutionContext(
                   static_cast<void*>(handle),
                   static_cast<void*>(execution_context));
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(execution_context);
 
@@ -363,14 +296,14 @@ hipdnnPluginStatus_t
                   static_cast<const void*>(device_buffers),
                   num_device_buffers);
 
-    return hipdnn_plugin::try_catch([&, api_name = __func__]() {
+    return try_catch([&, api_name = __func__]() {
         throw_if_null(handle);
         throw_if_null(execution_context);
         throw_if_null(device_buffers);
 
         // Simulate execution failure by throwing an exception
-        throw hipdnn_plugin::Hipdnn_plugin_exception(HIPDNN_PLUGIN_STATUS_INTERNAL_ERROR,
-                                                     "Simulated execution failure for testing");
+        throw Hipdnn_plugin_exception(HIPDNN_PLUGIN_STATUS_INTERNAL_ERROR,
+                                      "Simulated execution failure for testing");
 
         LOG_API_SUCCESS(api_name, "executed graph");
     });
