+++ tests/test_plugins/test_plugin_common.hpp
@@ -1,0 +1,12 @@
+// Copyright © Advanced Micro Devices, Inc., or its affiliates.
+// SPDX-License-Identifier: MIT
+
+struct hipdnnEnginePluginHandle
+{
+public:
+    virtual ~hipdnnEnginePluginHandle() = default;
+};
+
+struct hipdnnEnginePluginExecutionContext
+{
+};
